// mollys-movies-ui/stubs/api/model/movie-paginated-data.ts
// tslint:disable
/**
 * Public Molly's Movies API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { Movie } from './movie';

/**
 * 
 * @export
 * @interface MoviePaginatedData
 */
export interface MoviePaginatedData {
    /**
     * 
     * @type {number}
     * @memberof MoviePaginatedData
     */
    page?: number;
    /**
     * 
     * @type {number}
     * @memberof MoviePaginatedData
     */
    limit?: number;
    /**
     * 
     * @type {number}
     * @memberof MoviePaginatedData
     */
    count?: number;
    /**
     * 
     * @type {Array<Movie>}
     * @memberof MoviePaginatedData
     */
    data?: Array<Movie> | null;
}